git: move git plugins into overlays
