--- lua/plugins/git/init.lua
@@ -1,8 +1,0 @@
--- initialize key mappings
-require("plugins.git.keymap")
-
--- register plugin repos
-local repos = require("plugins.git.repos")
-require("plugins").register(repos)
-
-require("plugins.git.autocmd").setup()
